Changed the warning messages and corrected behavior when square uncertainty is neagitve. Previously, if it was within the tolerance (e.g. -1E-13 with cutoff of 1E-10), a True/False was being passed into d2DeltaF[True/False] which was  doing:
d2DeltaF[0,:] = 0 or d2DeltaF[1,:] = 0

instead of what I think it was trying to do of

d2DeltaF[d2DeltaF < 0.0] = 0.0

If any of the negative values of d2DeltaF are greater in magnitude than
the cutoff, a warning is issued saying how many entries violate the
cutoff and no entries are set to 0.0. This results in NaN's for
dDeltaF_ij, but such behavior should be left to the user to figure out
why they are getting them.