Start Coverage Improvements (Mbar.py)
This PR starts the coverage improvements for the package as a whole.
The focus here is on the mbar.py file as well as some initial ignores
for the CodeCov configs.

I also found and fixed 2 bugs with this:
- `svd` method missed a `@` operator (introduced in #334 by me)
- "BAR" init method failed due to dict return.

One outstanding question I have is the `computeCovarianceOfSums`
function. It's not used anywhere in the code and the last time it was
used was some time ago (I can go find the commit again if need be).
GitHub also does not seem to think anyone other than PyMBAR clones
use the function either.
If we want to keep this function around, is there a decent test which
can be written for it? The one I added simply runs it, not sure if its
accurate though.