Align namespaces to work if JAX is not installed
This sets all of the imports of JAX to align with the standard NumPy and Scipy imports so all the code continues functioning when JAX is not installed or has an issue installing/importing.

Adds a passthrough identity decorator so if `jax.jit` is unavailable, the decorator just returns the function, doing nothing.

This does create a few odd names, but they should be clear enough to function without issue.

Added a new CI variable to test mbar without JAX installed.

Because `use_jit` becomes `False`, the codepaths with the decorator may not be tested properly yet.