Make Black happy
Most of these were spaces around `**` operators.

`--exclude` doesn't really behave like most other program `--exclude` flags in that its not multi-mark and overwrites default settings and itself. Using the `--extend-exclude` so the `.gitignore` files are still respected and used a valid RegEx pattern to match the 2 files we are trying to ignore.