Update fes.py (BUG!)
This is a one-line code change to mitigate a bug that results in histogramming exiting with a key error.  For some reason, histogram bins that actually ARE populated are not correctly being added to the histogram_data dictionary. When the free energy for that bin is meant to be pulled, that bin doesn't exist in the dictionary and the KeyError is thrown. This fix assigns a default value of -1 to the bin_label if the key isn't present nstead of throwing this exception, and thus can allow the user to bypass the troubled bin(s) and still get a free energy surface. This fix does not find the root cause of the issue. I would be happy to provide the example which resulted in me finding the bug, but would require transferring many files.