Make JAX an optional dependency
This PR makes JAX an optional dependency at the pip level and raises a warning if JAX could not be imported to tell users it could be faster.

On the `pip` side, there will be 2 different instal options:

* `pip install pymbar` will install without JAX
* `pip install pymbar[jax]` will install with JAX.

Partial progress on #500. Until the [conda feedstock](https://github.com/conda-forge/pymbar-feedstock/issues/34) issue is resolved with a new recipe, I want to leave #500 open to keep documenting it.

Note: there is an outstanding TODO on the `README.md` file to fill in once we have an official name for the lite version on conda-forge.