After some testing, I found the speed gain from having pure static generated methods is negligible and my earlier testing was the fact I was re-instancing the solver, and thus re-JIT'ing everything each time a new MBAR was called (fixed in earlier commit).
After testing, here are the results:

Testing the timing of test_protocols test using static-generated methods as a relative baseline:
The test is 99% as fast on average with PyTree registration.
The test is 95% as fast on average without the PyTree registration.

So I've opted to use the JAX PyTree registration method and simplify the code substantially by moving all methods back into self methods.

Also updated the readme to reflect the new option.